fixed disableNbJavac issue, refactored server startup code into multiple files
