+++ vscode/src/configurations/handlers.ts
@@ -136,5 +136,5 @@ export const userdirHandler = (): string => {
 }
 
 export const isNbJavacDisabledHandler = (): boolean => {
-    return getConfigurationValue(configKeys.verbose, false);
+    return getConfigurationValue(configKeys.disableNbJavac, false);
 }
